split Controllers and moved Queue to server.ts
