+++ .eslintrc.json
@@ -21,5 +21,7 @@
 		"sourceType": "module"
 	},
 	"plugins": ["react", "@typescript-eslint"],
-	"rules": {}
+	"rules": {
+		"@typescript-eslint/interface-name-prefix": "off"
+	}
 }
