+++ src/controller/index.ts
@@ -1,0 +1,4 @@
+export * from "./screenerController";
+export * from "./studentController";
+export * from "./statisticsController";
+export * from "./queueController";
