+++ src/controller/screenerSocket.ts
@@ -1,5 +1,4 @@
-import ScreeningService from "../service/screeningService";
-import { io } from "../server";
+import { io, studentQueue } from "../server";
 
 enum screenerSocketEvents {
   LOGIN = "loginScreener",
@@ -13,11 +12,6 @@ enum screenerSocketActions {
 
 const SCREENER_CHANNEL = "screener";
 
-const screeningService = new ScreeningService();
-
-const subcriber = screeningService.myQueue.getClient().duplicate();
-subcriber.subscribe("queue");
-
 interface ScreenerInfo {
   firstname: string;
   lastname: string;
@@ -28,6 +22,9 @@ const allScreener: Map<string, string> = new Map([]);
 let onlineScreenerList: ScreenerInfo[] = [];
 
 const startScreenerSocket = (): void => {
+  const subcriber = studentQueue.getClient().duplicate();
+  subcriber.subscribe("queue");
+
   const addScreener = (screener: ScreenerInfo): void => {
     if (!onlineScreenerList.some((s) => s.email === screener.email)) {
       onlineScreenerList.push(screener);
@@ -58,7 +55,7 @@ const startScreenerSocket = (): void => {
 
       allScreener.set(socket.id, data.email);
 
-      const jobList = await screeningService.myQueue.listInfo();
+      const jobList = await studentQueue.listInfo();
       socket.join(SCREENER_CHANNEL);
       socket.emit(screenerSocketActions.UPDATE_QUEUE, jobList);
       addScreener(data);
@@ -82,7 +79,7 @@ const startScreenerSocket = (): void => {
   });
   subcriber.on("message", async () => {
     // screener is notified in everytime when the queue changes (we dont need to check what changed)
-    const jobList = await screeningService.myQueue.listInfo();
+    const jobList = await studentQueue.listInfo();
     io.sockets
       .in(SCREENER_CHANNEL)
       .emit(screenerSocketActions.UPDATE_QUEUE, jobList);
