+++ src/models/Screener.ts
