+++ src/models/StudentScreeningResult.ts
@@ -1,0 +1,9 @@
+export interface IStudentScreeningResult {
+  verified: boolean;
+  birthday?: Date;
+  commentScreener?: string;
+  knowscsfrom?: string;
+  subjects?: string;
+  feedback?: string;
+  screenerEmail: string;
+}
